Condition
Oddly enough, most people plan their vacations early. A young programmer has a certain budget and free time in a given season. Write a program that takes the budget and season as input, and outputs where the programmer will vacation and how much he will spend.
The budget determines the destination, and the season determines how much of the budget will be spent. If it's summer he'll vacation at a campground and winter at a hotel. If it's in Europe, regardless of the season he'll vacation in a hotel. Each campsite or hotel, depending on the destination, has its own price that corresponds to a certain percentage of the budget: 
- 100BGN or less - somewhere in Bulgaria
o Summer - 30% of the budget
o Winter - 70% of the budget
- At 1000BGN or less - somewhere in the Balkans
o Summer - 40% of the budget
o Winter - 80% of the budget
- At more than 1000BGN - somewhere in Europe
o When travelling in Europe, regardless of the season, you will spend 90% of the budget.
Input:
Login is read from the console and consists of two lines entered by the user:
- First line - Budget, a real number in the range [10.00...5000.00].
- Second line - One of two possible seasons: "summer" or "winter"
Output:
Two lines should be printed on the console.
- First line - "Somewhere in [destination]" between "Bulgaria", "Balkans" and "Europe"
- Second line - "{Vacation type} - {Amount spent}"
o Holiday can be between "Camp" and "Hotel"
o Amount must be rounded to the second decimal place.
Example:
Input:            Output:
50                 Somewhere in Bulgaria
summer        Camp - 15.00