Condition
Write a program that reads two integers (N1 and N2) and an operator to perform a mathematical operation on them. The possible operations are Addition(+), Subtraction(-), Multiplication(*), Division(/) and Modular division(%). When adding, subtracting and multiplying, the console should print the result and whether it is even or odd. In ordinary division, the result. In modular division, the remainder. It should be kept in mind that the divisor can be equal to 0(zero), and zero is not divisible. In this case a special message must be printed.
Input:
The 3 lines entered by the user are read from the console:
- N1 - integer in the range [0...40 000]
- N2 - integer in the range [0...40 000]
- Operator - one character with the following characters: "+", "-", "*", "/", "%"
Output:
Print one line on the console:
- If the operation is addition, subtraction or multiplication:
o "{N1} {operator} {N2} = {result} - {even/odd}"
- If the operation is division:
o "{N1} / {N2} = {result}" - the result is formated to the second decimal place
- If the operation is modular division: 
o "{N1} % {N2} = {result}"
- In case of division by 0(zero): 
o "Cannot divide {N1} by zero"
Example:
Input:        Output:
10             10 + 12 = 22 - even
12
+