Condition
A student must go to an exam at a certain time. He comes to the examination hall at the given arrival time. A student is considered to be on time if he/she arrives at the time of the exam or within half an hour before. If he arrives more than 30 minutes early, he is tardy. If he arrived after the exam time, he was late. Write a program that reads the exam time and arrival time and prints whether the student arrived on time, whether he was early or late, and by how many hours or minutes he was early or late.

Input:
The console reads 4 integers (one per line) entered by the user:
- The first line contains the time of the exam - an integer from 0 to 23.
- The second line contains the minute of the exam - an integer from 0 to 59.
- The third line contains the time of arrival - an integer from 0 to 23.
- The fourth line contains the minute of arrival, an integer from 0 to 59.
Output:
On the first line print:
- "Late" if the student arrives later than the exam time.
- "On time" if the student arrives exactly at the exam time or up to 30 minutes early.
- "Early" if the student arrives more than 30 minutes before the exam time.
If the student arrives at least one minute behind the exam time, print on the next line:
Example:
Input:        Output:
9               Late
30             20 minutes after the start
9
50