Condition
Write a program that calculates how much total money is in the account after making a certain number of contributions. On each line, you will get the amount you need to deposit into the account, until you receive a "NoMoreMoney " command. For each amount received, the console should display "Increase: " + the amount and add it to the account. If a number less than 0 is received the console should display "Invalid operation!" and the program should exit. When the program completes it should print "Total: " + the total amount in the account formatted to the second decimal place.