Condition
Write a program that calculates a student's average grade across all of his or her studies. On the first line you will get the student's name, and on each subsequent line his yearly grades. The student moves on to the next grade if his annual grade is greater than or equal to 4.00. If the student is suspended more than once, he is expelled and the program ends, printing the student's name and in which grade he is suspended.
 Upon successful completion of the 12th grade to print : 
"{student's name} graduated. Average grade: {average grade of entire course}"
In case the student is expelled from school, print:
"{student's name} has been excluded at {grade in which excluded} grade"
The value should be formatted to the second decimal place.