Condition
You're invited to a 30th birthday party, where the birthday boy treats you to a huge cake. However, he doesn't know how many pieces the guests can take from it. Your task is to write a program that calculates the number of pieces the guests have taken before it runs out. You will get the dimensions of the cake in centimeters (width and length - integers in the range [1...1000]) and then, on each line until you receive the command "STOP" or until the cake is finished, the number of pieces the guests take from it. The pieces are 1 cm square.
Print one of the following lines on the console:
- "{number of pieces} pieces are left." - if you reach STOP and there are pieces of cake left.
- "No more cake left! You need {number of pieces missing} pieces more."
Example:
Input:        Output:
10             No more cake left! You need 1 pieces more.
10
20
20
20
20
21