Condition
Gosho works in a restaurant and is responsible for loading the dishwasher at the end of the day.
Your task is to write a program that calculates whether a given amount of dishwasher detergent bottles purchased is enough to wash a certain amount of dishes. It is known that each bottle contains 750 ml. of detergent, for 1 dish it takes 5 ml. and for a saucepan 15 ml.  Assume that every third load of dishes, the dishwasher is filled only with pots and the other times with plates. Until you receive an "End" command you will continue to receive a count of dishes to be washed.
Input:
From the console the following are read:
- Number of bottles of detergent to be used for washing dishes - an integer in the range [1...10] 
On each subsequent line, until the "End" command is received or the amount of detergent is exhausted, the number of dishes to be washed - integer in the range [1...100]
Output:
In case the amount of detergent was sufficient to wash the dishes:
"Detergent was enough!"
"{number of clean dishes} dishes and {number of clean pots} pots were washed."
"Leftover detergent {amount of detergent left} ml."
In case the amount of detergent was not enough to wash the dishes:
"Not enough detergent, {amount of detergent not reached} ml. more necessary!"