Condition
Write a program that reads a hidden message in a sequence of characters. These are received one per line until the "End" command is received. The words are formed from the letters in the order they are read. Characters that are not Latin letters must be ignored. Words hidden in the stream are separated by a secret command of three letters - "c", "o" and "n". The first time one of these letters is received, it is marked as encountered, but not stored in the word. On each subsequent occurrence, it is stored normally in the word. Once all three characters of the command are present, the word and a space " " are printed. A new word is started, which likewise waits for the secret command to be printed.
Input:
A series of lines with one character on each is read from the console until the "End" command is received.
Output:
The console prints one line of each word after the secret command, followed by a space.