Condition
Write a program that checks all possible combinations of a pair of numbers in the interval of two given numbers. The output prints which in turn is the combination whose sum of the numbers equals a given magic number. If there is no combination satisfying the condition, a message is printed that none was found.
Input:
The input is read from the console and consists of three lines:
- First line - start of interval - integer in the range [1...999]
- Second line - end of interval - integer in the interval [greater than the first number...1000]
- Third line - magic number - integer in the interval [1...10000]
Output:
One line should be printed on the console according to the result:
- If a combination is found whose sum of numbers is equal to the magic number
o "Combination N:{sequence number} ({first number} + {second number} = {magic number})"
- If no combination satisfying the condition is found
o "{number of all combinations} combinations - neither equals {magic number}"
Example:
Input:        Output:
1               Combination N:4 (1 + 4 = 5)
10
5