Condition
Write a program that reads from the console two six-digit integers in the range 100000 to 300000. Always the first number entered will be less than the second. On the console, print on 1 line separated by a space all the numbers that are between the two numbers read from the console and satisfy the following condition:
- the sum of the digits in the even and odd positions are equal. If there are no numbers meeting the condition on the console no result is output.