Condition
Write a program that reads the integers in the range from the console until a "stop" command is received. Find the sum of all the entered prime numbers and the sum of all the entered non-prime numbers. Since negative numbers cannot be prime by math definition, if a negative number is input, display the following message "Number is negative.". In this case, the entered number is ignored and is not added to either of the two sums, and the program continues its execution, waiting for another number to be entered. 
The output should print the two sums found on two lines in the following format:
- "Sum of all prime numbers is: {prime numbers sum}"
- "Sum of all non prime numbers is: {non prime numbers sum}"