Condition
The Train the Trainers course is coming to an end and the final assessment is approaching. Your task is to help the jury that will evaluate the presentations by writing a program to calculate the average score of each presentation by a student, and finally the average score of all of them.
From the console on the first line, read the number of people on the jury n - an integer in the range [1...20]
Then on a separate line the name of the presentation is read - text
For each presentation, a new line reads n - the number of marks - a real number in the range [2.00...6.00]
After calculating the average score for a particular presentation, the console prints
 "{presentation name} - {average score}."
After receiving the "Finish" command, the console prints "Student's final grade is {average grade across all presentations}." and the program ends.
All grades must be formatted to the second decimal place.