Condition
Write a program that calculates the profit of an agency selling airline tickets. The agency sells airline tickets of different airlines. You will get information about the number of adult tickets sold and the number of child tickets sold. The net price of an adult ticket is determined by the airline and a child ticket is 70% cheaper. The agency adds a service charge to the net price of each ticket.  The Agency's final profit is 20% of the total price of all tickets.
Input:
5 lines are read from the console:
1.	Airline name - text
2.	Number of adult tickets - integer in the range [1...400]
3.	Number of child tickets - integer in the range [25...120]
4.	Net adult ticket price - real number in the range [100.0...1600.0]
5.	Service charge price - real number in the range [10.0...50.0]
Output:
Print the final sales profit on the console, in the following format:
- "The profit of your agency from {airline name} tickets is {profit for agency} lv."
The profit price should be formatted to the second digit after the decimal point.