Condition
Company-manufacturer of aluminum windows accepts orders for the manufacture and installation of the following price list for one piece. The company only accepts wholesale orders (over 10 pieces). Depending on the number of windows ordered, the company gives a different discount to its customers.
The company also offers delivery of its orders for 60 BGN.
Size         Single price      Discount
90X130 - 110 BGN	        Over 30 pieces - 5%  Over 60 pieces - 8%
100X150 - 140 BGN         Over 40 pieces - 6%  Over 80 pieces - 10%
130X180 - 190 BGN    	Over 20 pieces - 7%   Over 50 pieces - 12%
200X300 - 250 lv.	        Over 25 pieces - 9%  Over 50 pieces - 14%
If the order exceeds 99 pieces - an additional 4% discount will be applied to the final price (after the shipping cost, if any).
If order is less than 10 pcs "Invalid order" will be displayed on the console.
Input:
User enters 3 lines:
1.	Number of windows - integer in the range [0..1000];
2.	Type of windows - text "90X130" or "100X150" or "130X180" or "200X300";
3.	Method of receipt - text
- With delivery - "With delivery" 
- Without delivery - "Without delivery"
Output:
A single number is displayed - the value of the order, in the following format:
o "{Total order value} BGN"
Output to be formatted to two decimal places.