Condition
Pepi wants you to write a program to find out who is the best player in the World Championship. The information you get will be the player and how many goals he has scored. You are asked to print who is the player with the most goals and whether he has made a hat-trick. A hat-trick is when a player has scored 3 or more goals. If the player has scored 10 or more goals, the program should stop.
Input:
Two lines are read from the console until the "END" command is entered:
- Player name - text
- Number of goals scored - positive integer in the range [1 ... 10000]
Output:
2 lines to be printed on the console:
- On the first line:
            "{player name} is the best player!"
- On the second line :
o If the best player has made a hat trick:
                   "He has scored {number of goals} goals and made a hat-trick !!!"
o If the best player has not made a hat-trick:
                   "He has scored {number of goals} goals."