Condition
The technique in the Christmas decoration shop is spoiled. Items that contain even numbers in their barcode cannot be marked by cashiers. Your task is to write a program that generates all barcodes that do NOT contain even numbers in them.
Input:
- Two four-digit numbers that indicate the range of barcodes you need to change.
- First line - a four digit number - the beginning of the range. An integer in the range [1000...9999]
- Second line - four-digit number - the end of the range. Integer in the range [1000...9999]
Output:
The console should print all "barcodes" that do NOT contain an even digit in them, separated by a space.