Condition
The weather is warming up and clubs are running promising promotions. Write a program to calculate a club's revenue for the evening and whether the desired profit has been reached, knowing the following conditions: the price of a cocktail is the length of its name. If the price of an order is an odd number, there is a 25% discount on the order. 
Input:
From the console the following is read:
- On the first line - the desired club profit - a real number in the range [1.00... 15000.00]
A sequence of two lines until the "Party!" command is received or the desired win is reached:
o Cocktail name - text
o Number of cocktails to order - integer in the range [1... 50]
Output:
On the console, print one line first:
- When the command "Party!" is received:
 "We need {unavailable amount} leva more."
- On reaching the desired payout:
	"Target acquired."
Then print:
	"Club income - {club income} leva."
Money to be formatted to the second digit after the decimal.