Condition
Your task is to write a program that takes the name of a team and makes statistics for it. During a season, each team plays a certain number of football matches, and for each match the team is given points depending on the outcome of the match. There are three possible outcomes of a match:
W - The team is the winner and receives 3 points
D - The match ended without a winner and the team receives 1 point
L - The team lost the match and receives no points
Write a program that takes the name of a soccer team and outputs its statistics based on the games played this season. Its statistics should include the total number of points won this season, detailed statistics about the outcome of the games played, and the winning percentage this season. If the team has not played any games this season for any reason, a special message is displayed.
Input:
Two lines are read from the console:
- The name of the football team for which statistics are kept - Text
- Number of matches played in the current season - integer in the range [0... 100]
 A separate line is read for each match played:
o The result of the match played in one of the above formats - symbol with options 'W', 'D' and 'L'
Output:
Depending on whether the team has played matches in the current season, two types of output are displayed.
- If the team has not played any matches in the current season exactly one line is displayed in the following format:
o "{team name} hasn't played any games during this season."
- If the team has played one game or more, six lines are displayed in the following format:
o "{team name} has won {number of points won} points during this season"
o "Total stats:"
o "## W: {number of games won}"
o "## D: {number of games tied}" 
o "## L: {number of games lost}" 
o "Win rate: {percentage of games won}%"
The percentage of games won must be formatted to the second digit after the decimal point.