Condition
Ivan has invented a new game to compete with his friends. Your task is to write a program for the game. Each player writes their name, then for each letter of their name they write an integer, if the integer matches the ASCII value of that letter the player gets 10 points, otherwise they only get 2 points. The winner is the player with the most points at the end of the game. In case two players have the same number of points, the one who reached the score second wins.
Input:
One line is read until the "Stop" command is received:
- Player name of length n - text
For each player, n number of lines are read:
- number - integer in the range[0...127] 
Output:
Print one line in the following format:
- "The winner is {winner's name} with {winner's points} points!"