Condition
The teaching team of SoftUni organized a pool work day to celebrate the arrival of summer. Your task is to write a program that calculates how much should be paid. For each person on the team, an entrance fee must be paid. You should keep in mind that one umbrella is enough for two people. It is known that only 75% of the team want sunbeds. When calculating the number of umbrellas and sun loungers, round the number to the larger whole number.
Input:
4 numbers are read from the console:
1.	First line - number of people. Integer in the range [1...100]
2.	Second line - entrance fee. Real number in the range [0.00...50.00]
3.	Third line - price one per lounger. Real number in the range [0.00...50.00]
4.	Fourth row - price per umbrella. Real number in the range [0.00...50.00] 
Output:
"{amount to cover costs} lv." 
Format the result to the second digit after the decimal point.