Condition
Pesho decides that he wants to do some renovations at home. His task is to paint the walls in the living room, knowing the height and width of one wall. Pesho's living room has 4 walls of the same size, a certain percentage of which are taken up by windows and doors that will not be painted. He is not sure if he will be able to do it in one go, so he asks you to help him calculate whether he will have any more work to do the next day and, if so, how many square metres he has to finish and, if he is able to paint the living room, how much paint he has left (it should be borne in mind that one litre of paint paints one square metre of wall). 
Input:
The following lines are read from the console:
1.	Wall height - integer [0... 100]
2.	Wall width - integer [0... 100]
3.	Percentage of total wall area that will not be painted - integer [5... 95]
On the following lines until the command "Tired!" is received or until all walls are painted, one number is read:
- Liters of paint - integer [0...100]:
Note: The area to be painted is to be rounded up to the nearest whole number.
Output:
Print one of the following lines on the console:
- On receipt of the "Tired!" command:
"{square meters} quadratic m left." 
{square meters} is the surface area left to paint.
- If there is paint left over:
"All walls are painted and you have {litres of paint} l
 paint left!" 
- If after painting all the walls, there is no paint left:
"All walls are painted! Great job, Pesho!"