Condition
For Laura, words hold great power. She asks you to come up with an algorithm to find the "strongest" word. Until you receive an "End of words" command, words will be read from the console. To find the strength of each one, you need to find the sum of the ASCII values of the characters that make up the word. If it starts with a vowel letter - 'a', 'e', 'i', 'o', 'u', 'y' (or their equivalent capital letters), the resulting sum should be multiplied by the length of the word, otherwise, divide by the length and round down to the nearest whole number.
Input:
One line at a time is read from the console until the "End of words" command is received:
- word - text 
Output:
When the program is finished, the word with the "most power" is printed on one line:
- "The most powerful word is {the word with the most power} - {the power of the word}"