Condition
In order for cats to be healthy, their diet must follow a certain diet. Write a program that calculates a cat's percentage distribution of macronutrients in food for a day and calculates how many average calories one gram of that food provides. The macronutrients are: fat, protein, and carbohydrate.
You have the total number of calories a cat should take in for one day.
It is known that:
- 1 gram of fat = 9 calories
- 1 gram protein = 4 calories
- 1 gram of carbohydrate = 4 calories
To figure out how many calories one gram of food gives your cat, you will need to do calculations with the actual weight of the food, since it contains water. You need to calculate the grams of fat, protein and carbohydrates. Their sum gives the total weight of the food and from this we need to subtract the percentages of water. 
Input:
5 lines are read from the console:
- Fat percentage - integer in the range [0...100]
- Protein percentage - integer in the range [0...100]
- Carbohydrate percentage - integer in the range [0...100]
- Total calories - integer in the range [1000...15000]
- Percentage of water content - integer in the range [0...100]
Clarification: when doing such calculations with percentages there is a big chance that the result is not a whole number!
Output:
1 line is printed on the console:
- "{calories}"
The result must be formatted to the fourth decimal place.