Condition
Like a true fan, Pepi decided to prepare for the World Cup by buying a team of his favorite football team. The shop where he shopped offered t-shirts, shorts, socks and boots. It was known that the price of the shorts was 75% of the price of the t-shirts and the price of the socks was 20% of the price of the shorts. Buttons cost twice as much as the T-shirt and shorts combined. Since Peppy regularly shops from this store, he has a discount card worth 15% of the total purchase. If Pepi's bill is greater than or equal to a certain amount, he receives a gift - an exact replica of the World Ball. Write a program that calculates whether Pepi has won the ball. 
Input:
The input is read from the console and contains exactly 2 lines:
- On the first line - the price of the jersey - a real number in the range [1.00 ... 1000.00]
- On the second line - the amount you must reach to win the ball - a real number in the range [100.00 ... 10 000.00]
Output:
Two lines are printed on the console:
- If the ball IS won:
o "Yes, he will earn the world-cup replica ball!"
o "His sum is {sum} lv."
- If the ball IS NOT won:
o "No, he will not earn the world-cup replica ball."
o "He needs {not enough money} lv. more."
The result should be formatted to the second decimal place.