Condition
Courier company delivers parcels throughout the country. 
For standard service, the delivery time is 3 working days and the company calculates the prices under the following conditions:
- For parcels lighter than 1 kg - 3 cents per kilometer.
- From 1 kg to 10 kg - 5 cents per kilometer.
- From 10 kg incl. up to 40 kg - 10 cents per kilometer.
- From 40 kg incl. to 90 kg - 15 cents per km.
- From 90 kg incl. to 150 kg - 20 cents per kilometre.
For express service, the company will perform the service within 24 hours, charging a surcharge per kilometre as follows: 
- For parcels lighter than 1 kg - per kg at 80% of the respective price per kilometer
- From 1 kg to 10 kg - per kg at 40% of the respective price per km
- From 10 kg up to and including 40 kg - per kg at 5 % of the relevant price per kilometre
- From 40 kg incl. up to 90 kg - per kg at 2 % of the relevant price per kilometre
- From 90 kg incl. up to 150 kg - per kilogram at 1 % of the relevant price per kilometre
Write a program to calculate at a given distance in km. , the weight of the shipment and the type of service, what will be the cost of delivery of a shipment.
Login
The input is read from the console and contains 3 lines:
1.	Shipment weight in kg - a real number in the range [0.01 ... 150.00]
2.	Type of service - text with the following options.
3.	Distance in kilometres - an integer in the range [1 ... 1000]
Output
Print one line on the console:
"The delivery of your shipment with weight of {weight} kg. would cost {price} lv."
- Weight to be rounded to the third decimal place
- Price to be rounded to the second decimal place