Condition
Write a program to compute statistics for exam grades. At the beginning, the program gets the number of students who took the exam and, for each student, his or her grade. At the end, the program should print the percentage of students with scores between 2.00 and 2.99, between 3.00 and 3.99, between 4.00 and 4.99, 5.00 or more. Also the average grade on the exam.
Input:
From the console read:
- On the first line - the number of students who took the exam - an integer in the range [1...1000]
- For each student on a separate line - the exam grade - a real number in the range [2.00...6.00]
Output:
Print 5 lines on the console that contain the following information:
Line 1 - "Top students: {percentage of students with a grade of 5.00 or higher}%"
Line 2 - "Between 4.00 and 4.99: {between 4.00 and 4.99 inclusive}%"
Row 3 - "Between 3.00 and 3.99: {between 3.00 and 3.99 inclusive}%"
Line 4 - "Fail: {less than 3.00}%"
Line 5 - "Average: {average success}"
All numbers must be formatted to the second decimal place.