Condition
Tour company sells out final tour packages for the end of the year. Write a program to calculate the profit from the sale and whether or not all the packages were sold, knowing the original number of tours and their prices. The company offers two types of excursions - sea ("sea") for 680 leva and mountain ("mountain") for 499 leva. When choosing a package, the corresponding price is added to the total amount. If you choose a package that is no longer available (the number is 0), it should not be added to the total price.
Input:
2 lines are initially read from the console:
1.	Number of sea excursions - integer in the range [1... 500]
2.	Number of trips to the mountains - integer in the range [1... 500]
Then one line is read at a time until a "Stop" command is received or until the company has sold all packages:
- Package name - text with options "sea" or "mountain"
Output:
1 or 2 lines are printed to the console, as appropriate: 
- If the company has managed to sell all packages:
"Good job! Everything is sold." 
It always prints:
"Profit: {profit from sales} leva."